requirements.txt
PyYAML>=3.13
jmespath>=0.9.3
pyinotify==0.9.6
